refactor: Refactor move dateformattor func to lib/common.ts
